+++ runtime/src/bank.rs
@@ -1867,6 +1867,26 @@ pub fn goto_end_of_slot(bank: &mut Bank) {
     }
 }
 
+// This protects from memory exhaustion assuming being used only once when starting validator
+pub fn deserialize_for_snapshot<R, T>(reader: R) -> bincode::Result<T>
+where
+    R: std::io::Read,
+    T: serde::de::DeserializeOwned,
+{
+    let mem_info = sys_info::mem_info();
+
+    if let Ok(mem_info) = mem_info {
+        // Because we're are about to restore full process state from file,
+        // allow use of memory as much as possible while maintaining 20% memory available so as not to trigger our earlyoom
+        let mem_limit = mem_info.total.saturating_sub(mem_info.total / 5);
+        //let mem_limit = mem_info.avail.saturating_sub(mem_info.total / 5);
+        //eprintln!("use at most {} {:#?}", mem_limit, mem_info);
+        bincode::config().limit(mem_limit).deserialize_from(reader)
+    } else {
+        bincode::config().deserialize_from(reader)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1876,7 +1896,6 @@ mod tests {
         genesis_utils::{
             create_genesis_config_with_leader, GenesisConfigInfo, BOOTSTRAP_LEADER_LAMPORTS,
         },
-        serde_utils::deserialize_for_snapshot,
         status_cache::MAX_CACHE_ENTRIES,
     };
     use bincode::{serialize_into, serialized_size};
